+++ models/models.go
@@ -9,7 +9,7 @@ import (
 )
 
 type Author struct {
-	ID   int64  `json:"id"`
+	ID   int32  `json:"id"`
 	Name string `json:"name"`
 	Bio  string `json:"bio"`
 }
