+++ models/query.sql.go
@@ -36,7 +36,7 @@ DELETE FROM authors
 WHERE id = $1
 `
 
-func (q *Queries) DeleteAuthor(ctx context.Context, id int64) error {
+func (q *Queries) DeleteAuthor(ctx context.Context, id int32) error {
 	_, err := q.db.ExecContext(ctx, deleteAuthor, id)
 	return err
 }
@@ -46,7 +46,7 @@ SELECT id, name, bio FROM authors
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetAuthor(ctx context.Context, id int64) (Author, error) {
+func (q *Queries) GetAuthor(ctx context.Context, id int32) (Author, error) {
 	row := q.db.QueryRowContext(ctx, getAuthor, id)
 	var i Author
 	err := row.Scan(&i.ID, &i.Name, &i.Bio)
@@ -71,7 +71,6 @@ type GetBookDetailsByIDRow struct {
 	AuthorID sql.NullInt32 `json:"author_id"`
 }
 
-// Example of join query emabeded
 func (q *Queries) GetBookDetailsByID(ctx context.Context, id int64) (GetBookDetailsByIDRow, error) {
 	row := q.db.QueryRowContext(ctx, getBookDetailsByID, id)
 	var i GetBookDetailsByIDRow
@@ -122,7 +121,7 @@ WHERE id = $1
 `
 
 type UpdateAuthorParams struct {
-	ID   int64  `json:"id"`
+	ID   int32  `json:"id"`
 	Name string `json:"name"`
 	Bio  string `json:"bio"`
 }
